# optimize compare_no_case for str