@Geal 2018-02-10
that's a nice one, thanks :)

I was very annoyed about this, and that solution is nice and clean